requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.10
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
numpy==2.1.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
prov==2.0.1
pydot==3.0.2
PyMuPDF==1.24.12
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3


